Formatting fixes

+++ src/handlers/ControlHandler.java
@@ -45,62 +45,62 @@ public class ControlHandler {
         switch(input) {
             case UP: {
                 robot.keyPress(KeyEvent.VK_UP);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_UP);
                 break;
             }
             case DOWN: {
                 robot.keyPress(KeyEvent.VK_DOWN);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_DOWN);
                 break;
             }
             
             case LEFT: {
                 robot.keyPress(KeyEvent.VK_LEFT);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_LEFT);
                 break;
             }
             case RIGHT: {
                 robot.keyPress(KeyEvent.VK_RIGHT);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_RIGHT);
                 break;
             }
             case SQUARE: {
                 robot.keyPress(KeyEvent.VK_A);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_A);
                 break;
             }
             case CIRCLE: {
                 robot.keyPress(KeyEvent.VK_X);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_X);
                 break;
             }
             case CROSS: {
                 robot.keyPress(KeyEvent.VK_Z);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_Z);
                 break;
             }
             case TRIANGLE: {
                 robot.keyPress(KeyEvent.VK_S);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_S);
                 break;
             }
             case START: {
                 robot.keyPress(KeyEvent.VK_SPACE);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_SPACE);
                 break;
             }
             case SELECT: {
                 robot.keyPress(KeyEvent.VK_V);
-                Thread.sleep((long) (20));
+                Thread.sleep((20));
                 robot.keyRelease(KeyEvent.VK_V);
                 break;
             }
